Rename doctor components
